+++ src/main/java/com/smoky/bassshakertelemetry/client/ClientInit.java
@@ -13,6 +13,7 @@ public final class ClientInit {
 
     public static void init() {
         MinecraftForge.EVENT_BUS.register(new TelemetryEventHandler());
+        MinecraftForge.EVENT_BUS.register(new SoundHapticsHandler());
 
         ModLoadingContext.get().registerExtensionPoint(
                 ConfigScreenHandler.ConfigScreenFactory.class,